7/15 Drivetrain Written
OI , line 66 was changed to be a
static variable to allow access for
drivetraincommand.
Robot added subsystems for drivetrain.
Robot map added values for frontLeft, backLeft, frontRight, and backRight. Midright and midLeft were removed from speedcontroller group.
Closevalve extends instantcommand instead of command.
Drivetrain deleted and rewritten as a subsystem.
DrivetrainCommand is a command class for drivetrain, that describes requirements for the drivetrain.
Drivetrain was written as a subsystem to set the motors for the differential drive.